pyproject.toml
[tool.poetry]
name = "ai-sdlc-executor"
version = "0.1.0"
description = "AI-driven SDLC Orchestrator - Executor Agent following strict workflow protocol"
authors = ["AI Team <team@company.com>"]
readme = "README.md"
packages = [{include = "executor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# CLI Framework
typer = {extras = ["all"], version = "^0.12.0"}
rich = "^13.7.0"

# MCP Protocol
mcp = "^1.0.0"

# API Clients (for custom MCP servers)
requests = "^2.31.0"
httpx = "^0.27.0"

# Data Validation
pydantic = "^2.8.0"
pydantic-settings = "^2.3.0"

# LLM
anthropic = "^0.34.0"

# Utilities
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"

# HTML/Markdown parsing (for Confluence cleaning)
beautifulsoup4 = "^4.12.0"
html2text = "^2024.2.26"
markdown = "^3.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.5.0"
mypy = "^1.10.0"

[tool.poetry.scripts]
orchestrator = "executor.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
